Create wrapper function for api_tools + Re-factor
- Re-factor code and create a wrapper function to access private api_calls.
- Fix wrong function name
